some renaming cos confused and tests pass..
