+++ simulation/point_geom.go
