+++ simulation/train_types.go
